requirements.txt
torch~=2.4.0+cu124
python-dotenv~=1.0.1
openai~=1.42.0
requests~=2.32.3
diffusers~=0.30.3